Por todos lados, cambio la matriz del grafo en niveles por otra con sus valores en doubles, porque hay un bug muy feo en la comparacion con infinito en los algoritmos matriciales

+++ problema2/src/Dantzig.h
@@ -19,7 +19,7 @@ public:
     /**
      * Metodo para resolver camino minimo usando Dantzig.
      */
-    virtual std::vector<std::vector<int>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n);
+    virtual std::vector<std::vector<double>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n);
 
 };
