+++ problema2/src/DijkstraPQ.h
@@ -20,16 +20,16 @@ public:
      * de camino minimo. Las clases que implementen esta interfaz
      * deben implementar este metodo.
      */
-    virtual std::vector<std::vector<int>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n);
+    virtual std::vector<std::vector<double>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n);
 private:
 
     /**
      * Funcion auxiliar que corre Dijsktra con priority queue sobre un vertice.
      * Se usa para correr Dijkstra secuencialmente sobre todos los vertices del grafo.
      */
-    void DijkstraPQAux(std::vector<int> &min_path_len,
+    void DijkstraPQAux(std::vector<double> &min_path_len,
         int vertex,
-        std::vector<std::vector<int>>& distancias);
+        std::vector<std::vector<double>>& distancias);
 };
 
 #endif
