+++ problema2/src/ShortestPath.h
@@ -21,14 +21,14 @@ public:
      * de camino minimo. Las clases que implementen esta interfaz
      * deben implementar este metodo.
      */
-    virtual std::vector<std::vector<int>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n)=0;
+    virtual std::vector<std::vector<double>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n)=0;
 
 protected:
     /**
      * Metodo auxiliar para armar el grafo en niveles en base a las rutas existentes
      * y los costos de la nafta en cada ciudad.
      */
-    std::vector<std::vector<int>> armarGrafoEnNiveles(std::vector<Ruta>& rutas,
+    std::vector<std::vector<double>> armarGrafoEnNiveles(std::vector<Ruta>& rutas,
         std::vector<int> costos,
         int n);
